+++ include/data_types/hashtable.h
@@ -18,7 +18,7 @@ typedef t_hashtable;
 
 struct s_bucket_data
 {
-	int key;
+	u64 key;
 	void* value;
 }
 typedef t_bucket_data;
@@ -27,9 +27,9 @@ t_hashtable* hashtable_init(u32 init_size);
 t_hashtable* dc_hashtable_init(u32 init_size,void (*data_destructor)(void*));
 void hashtable_free(t_hashtable* hashtable);
 void hashtable_dispose(t_hashtable* hashtable);
-void* hashtable_get(t_hashtable* hashtable,u32 key);
-void* hashtable_remove(t_hashtable* hashtable,u32 key);
-void hashtable_put(t_hashtable* hashtable,u32 key,void* value);
+void* hashtable_get(t_hashtable* hashtable,u64 key);
+void* hashtable_remove(t_hashtable* hashtable,u64 key);
+void hashtable_put(t_hashtable* hashtable,u64 key,void* value);
 void hashtable_put_str(t_hashtable* hashtable,unsigned char* key,char* value);
 t_hashtable* hashtable_clone_map(t_hashtable* map,u32 data_size);
